+++ Tiles/src/Tiles/InfiniteGrid.cpp
@@ -14,9 +14,9 @@ InfiniteGrid::~InfiniteGrid()
 void InfiniteGrid::InitializeGeometry() {
     GridData data;
 
-    m_Attributes.AddBuffer("a_Position", 0, data.positions.data(), sizeof(float), data.positions.size(), 3, false);    
-    m_Attributes.AddIndices(data.indices.data(), data.indices.size());
-    m_VertexArray.ApplyAttributes(m_Attributes);
+    m_Attributes.AddBuffer("a_Position", 0, data.positions.data(), sizeof(float), (uint32_t)data.positions.size(), 3, false);    
+    m_Attributes.AddIndices(data.indices.data(), (uint32_t)data.indices.size());
+    m_VertexArray.ApplyAttributes(m_Attributes); 
 }
 
 void InfiniteGrid::Draw(GL::ShaderProgram& shader)
