+++ Tiles/src/Tiles/TileObject.cpp
@@ -16,11 +16,11 @@ void TileObject::InitializeGeometry() {
     TileData tileData;
 
     // Per vertex
-    m_Attributes.AddBuffer("a_Position", 0, tileData.positions.data(), sizeof(float), tileData.positions.size(), 3, false);
-    m_Attributes.AddBuffer("a_Normal", 1, tileData.normals.data(), sizeof(float), tileData.normals.size(), 3, false);
-    m_Attributes.AddBuffer("a_TexCoord", 2, tileData.texCoords.data(), sizeof(float), tileData.texCoords.size(), 2, false);
+    m_Attributes.AddBuffer("a_Position", 0, tileData.positions.data(), sizeof(float), (uint32_t)tileData.positions.size(), 3, false);
+    m_Attributes.AddBuffer("a_Normal", 1, tileData.normals.data(), sizeof(float), (uint32_t)tileData.normals.size(), 3, false);
+    m_Attributes.AddBuffer("a_TexCoord", 2, tileData.texCoords.data(), sizeof(float), (uint32_t)tileData.texCoords.size(), 2, false);
 
-    m_Attributes.AddIndices(tileData.indices.data(), tileData.indices.size());
+    m_Attributes.AddIndices(tileData.indices.data(), (uint32_t)tileData.indices.size());
 
     m_VertexArray.ApplyAttributes(m_Attributes);
